+++ CHANGELOG.md
@@ -327,6 +327,10 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 - $Cache(x, y)$ means BST with length y starting at x
 - $Cache(x, 0)$ is ${ nil }$
 
+## 0098 - Validate Binary Search Tree
+
+- Recursion
+
 ## 0415 - Add strings
 
 - Full adder